Fixed set wifi network credentials script

+++ set-wifi-network-credentials.sh
@@ -17,8 +17,13 @@ fi
 echo "Setting WiFi network credentials..."
 echo "  WiFi network name: $WIFI_NAME"
 
-cp wifi-network-name.security wifi-network-name-previous.security
-cp wifi-network-password.security wifi-network-password-previous.security
+if [ -f wifi-network-name.security ]; then
+  cp wifi-network-name.security wifi-network-name-previous.security
+fi
+
+if [ -f wifi-network-password.security ]; then
+  cp wifi-network-password.security wifi-network-password-previous.security
+fi
 
 echo $WIFI_NAME > wifi-network-name.security
 echo $WIFI_PASSWORD > wifi-network-password.security
